Finished test 2
Test 2 is finished with some problems when it comes to showing the data on screen. while the calls to the server looks fine and it is able to add items and fetch everything from the db its not able to show it, maybe its something specific with the getAll call.
Now I will proceed to work on the test 3 that uses the mysql1 library.